+++ src/fonduer/parser/lingual_parser/__init__.py
@@ -1,0 +1,5 @@
+from fonduer.parser.lingual_parser.lingual_parser import LingualParser
+from fonduer.parser.lingual_parser.simple_parser import SimpleParser
+from fonduer.parser.lingual_parser.spacy_parser import SpacyParser
+
+__all__ = ["LingualParser", "SpacyParser", "SimpleParser"]