+++ tests/parser/test_spacy_parser.py
@@ -2,49 +2,44 @@ import logging
 
 import pytest
 
-from fonduer.parser.models import Sentence
-from fonduer.parser.spacy_parser import (
-    Spacy,
+from fonduer.parser.lingual_parser.spacy_parser import (
+    SpacyParser,
     TokenPreservingTokenizer,
     set_custom_boundary,
 )
+from fonduer.parser.models import Sentence
 
 
 def test_spacy_support(caplog):
     caplog.set_level(logging.INFO)
 
     # Supported language
-    lingual_parser = Spacy("en")
+    lingual_parser = SpacyParser("en")
     assert lingual_parser.has_tokenizer_support()
     assert lingual_parser.has_NLP_support()
 
     # Alpha-supported language
-    lingual_parser = Spacy("ja")
+    lingual_parser = SpacyParser("ja")
     assert lingual_parser.has_tokenizer_support()
     assert not lingual_parser.has_NLP_support()
 
     # Non supported language
-    lingual_parser = Spacy("non-supported-lang")
+    lingual_parser = SpacyParser("non-supported-lang")
     assert not lingual_parser.has_tokenizer_support()
     assert not lingual_parser.has_NLP_support()
 
     # Language not specified
     with pytest.raises(TypeError):
-        lingual_parser = Spacy()
+        lingual_parser = SpacyParser()
 
 
 def test_spacy_split_sentences(caplog):
     caplog.set_level(logging.INFO)
 
-    lingual_parser = Spacy("en")
+    lingual_parser = SpacyParser("en")
     tokenize_and_split_sentences = lingual_parser.split_sentences
     text = "This is a text. This is another text."
 
-    iterator = tokenize_and_split_sentences(text)
-    with pytest.raises(AttributeError):
-        next(iterator)
-
-    lingual_parser.load_lang_model()
     iterator = tokenize_and_split_sentences(text)
     assert len(list(iterator)) == 2
 
@@ -52,8 +47,7 @@ def test_spacy_split_sentences(caplog):
 def test_split_sentences_by_char_limit(caplog):
     caplog.set_level(logging.INFO)
 
-    lingual_parser = Spacy("en")
-    lingual_parser.load_lang_model()
+    lingual_parser = SpacyParser("en")
 
     text = "This is a text. This is another text."
     all_sentences = [